Cross-route image animation, Step 2: Slick animation!
Here we (finally) use the custom "crossfade" transition functions that I committed too early, in #04f1599.

Note the use of the `key` parameter when invoking these. `receive` and `send` use this value to match animation targets across routes and, as such, it's basically non-optional.

The result isn't interruptible, so over-frantic interactions can cause weird, if temporary, visual glitches, but if you're already committed to CSS animations*, this pattern's definitely worth a shot!**

*As opposed to JS-based animations from the likes of [`GSAP`](https://github.com/greensock/GreenSock-JS) or [`elm-ui`](https://github.com/mdgriffith/elm-ui))

**Fwiw, learning Svelte has felt similar to using Ruby after growing up with Objective-C: fun, light, targeted, powerful. But unlike Ruby (and it's not-so-light friend, rails), Svelte has the additional benefits of compiler feedback and near-zero translation between the code one writes and the code one steps through in-browser. Love it!