Adding more animation effects
Note how we're animating size, color, and position (and, cheating the mind with a quick crossfade, the text _content_ as well), and the tag doesn't even have to stay consistent! We go from `<h2>` to `<p>` to `<blockquote>`, all without skipping a beat. This is probably just how native CSS animations work, but I don't currently care…I just think it's cool that this can be described so succinctly in Svelte!